# ci/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "VideoOrganizer API",
    "description": "Self-hosted media library manager \u2014 videos, tags, custom properties, background workers (thumbnail sprites \u002B Md5 dedup), and bulk import.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost/"
    }
  ],
  "paths": {
...
    "/api/videos": {
      "get": {
        "tags": [
          "VideoOrganizer.API"
        ],
        "operationId": "GetVideos",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VideoDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "VideoOrganizer.API"
        ],
        "operationId": "CreateVideo",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Video"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Video"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/videos/{id}": {
      "get": {
        "tags": [
          "VideoOrganizer.API"
        ],
        "operationId": "GetVideoById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "VideoOrganizer.API"
        ],
        "operationId": "UpdateVideo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVideoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "VideoOrganizer.API"
        ],
        "operationId": "DeleteVideo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/videos/{id}/properties": {
      "put": {
        "tags": [
          "VideoOrganizer.API"
        ],
        "operationId": "SetVideoProperties",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPropertyValuesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/tags/{id}/properties": {
      "put": {
        "tags": [
          "Tags"
        ],
        "operationId": "SetTagProperties",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPropertyValuesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/properties": {
      "get": {
        "tags": [
          "Properties"
        ],
        "operationId": "ListProperties",
        "parameters": [
          {
            "name": "tagGroupId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PropertyDefinitionDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Properties"
        ],
        "operationId": "CreateProperty",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePropertyDefinitionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyDefinitionDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/properties/{id}": {
      "put": {
        "tags": [
          "Properties"
        ],
        "operationId": "UpdateProperty",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePropertyDefinitionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Properties"
        ],
        "operationId": "DeleteProperty",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/playlists/random": {
      "post": {
        "tags": [
          "Playlists"
        ],
        "operationId": "CreateRandomPlaylist",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/PlaylistFilterRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/playlists/even": {
      "post": {
        "tags": [
          "Playlists"
        ],
        "operationId": "CreateEvenDistributionPlaylist",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/PlaylistFilterRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/playlists/{id}": {
      "get": {
        "tags": [
          "Playlists"
        ],
        "operationId": "GetPlaylist",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/playlists/{playlistId}/navigation/{videoId}": {
      "get": {
        "tags": [
          "Playlists"
        ],
        "operationId": "GetPlaylistNavigation",
        "parameters": [
          {
            "name": "playlistId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "videoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistNavigationDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PropertyDefinitionDto": {
        "required": [
          "id",
          "name",
          "dataType",
          "scope",
          "tagGroupId",
          "required",
          "sortOrder",
          "notes"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "dataType": {
            "$ref": "#/components/schemas/PropertyDataTypeDto"
          },
          "scope": {
            "$ref": "#/components/schemas/PropertyScopeDto"
          },
          "tagGroupId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "required": {
            "type": "boolean"
          },
          "sortOrder": {
            "type": "integer",
            "format": "int32"
          },
          "notes": {
            "type": "string"
          }
        }
      },
...
      "SetPropertyValuesRequest": {
        "required": [
          "values"
        ],
        "type": "object",
        "properties": {
          "values": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PropertyValueWrite"
            }
          }
        }
      },
...
      "Video": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fileName": {
            "type": "string"
          },
          "filePath": {
            "type": "string"
          },
          "md5": {
            "type": [
              "null",
              "string"
            ]
          },
          "md5Failed": {
            "type": "boolean"
          },
          "md5FailedError": {
            "type": [
              "null",
              "string"
            ]
          },
          "thumbnailsFailed": {
            "type": "boolean"
          },
          "thumbnailsFailedError": {
            "type": [
              "null",
              "string"
            ]
          },
          "thumbnailsGenerated": {
            "type": "boolean"
          },
          "importJobId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "fileSize": {
            "type": "integer",
            "format": "int64"
          },
          "duration": {
            "pattern": "^-?(\\d\u002B\\.)?\\d{2}:\\d{2}:\\d{2}(\\.\\d{1,7})?$",
            "type": "string"
          },
          "height": {
            "type": "integer",
            "format": "int32"
          },
          "width": {
            "type": "integer",
            "format": "int32"
          },
          "videoDimensionFormat": {
            "$ref": "#/components/schemas/DomainVideoDimensionFormat"
          },
          "videoCodec": {
            "$ref": "#/components/schemas/DomainVideoCodec"
          },
          "bitrate": {
            "type": "integer",
            "format": "int64"
          },
          "frameRate": {
            "type": "number",
            "format": "double"
          },
          "pixelFormat": {
            "type": [
              "null",
              "string"
            ]
          },
          "ratio": {
            "type": [
              "null",
              "string"
            ]
          },
          "creationTime": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "videoStreamCount": {
            "type": "integer",
            "format": "int32"
          },
          "audioStreamCount": {
            "type": "integer",
            "format": "int32"
          },
          "ingestDate": {
            "type": "string",
            "format": "date-time"
          },
          "cameraType": {
            "$ref": "#/components/schemas/DomainCameraTypes"
          },
          "videoQuality": {
            "$ref": "#/components/schemas/DomainVideoQuality"
          },
          "watchCount": {
            "type": "integer",
            "format": "int32"
          },
          "notes": {
            "type": "string"
          },
          "needsReview": {
            "type": "boolean"
          },
          "playbackIssue": {
            "type": "boolean"
          },
          "markedForDeletion": {
            "type": "boolean"
          },
          "isFavorite": {
            "type": "boolean"
          },
          "parentVideoId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "clipStartSeconds": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "clipEndSeconds": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "isClip": {
            "type": "boolean"
          },
          "isExportedClip": {
            "type": "boolean"
          },
          "isEdited": {
            "type": "boolean"
          },
          "clipExported": {
            "type": "boolean"
          },
          "exportedToVideoId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "ocrScannedThroughSeconds": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "chapterMarkers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChapterMarker"
            }
          },
          "videoBlocks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VideoBlock"
            }
          },
          "ocrTextLines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OcrTextLine"
            }
          },
          "videoTags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VideoTag"
            }
          },
          "propertyValues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VideoPropertyValue"
            }
          }
        }
      },
...
      "VideoCodec": {
        "enum": [
          "h264",
          "h265",
          "hevc",
          "notChecked",
          "other"
        ],
        "type": "string"
      },
...
      "VideoDto": {
        "required": [
          "id",
          "fileName",
          "filePath",
          "md5",
          "md5Failed",
          "md5FailedError",
          "thumbnailsFailed",
          "thumbnailsFailedError",
          "thumbnailsGenerated",
          "importJobId",
          "fileSize",
          "duration",
          "height",
          "width",
          "videoDimensionFormat",
          "videoCodec",
          "bitrate",
          "frameRate",
          "pixelFormat",
          "ratio",
          "creationTime",
          "videoStreamCount",
          "audioStreamCount",
          "ingestDate",
          "cameraType",
          "videoQuality",
          "watchCount",
          "notes",
          "needsReview",
          "playbackIssue",
          "markedForDeletion",
          "isFavorite",
          "parentVideoId",
          "clipStartSeconds",
          "clipEndSeconds",
          "isClip",
          "isExportedClip",
          "isEdited",
          "chapterMarkers",
          "videoBlocks",
          "tags",
          "properties"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fileName": {
            "type": "string"
          },
          "filePath": {
            "type": "string"
          },
          "md5": {
            "type": [
              "null",
              "string"
            ]
          },
          "md5Failed": {
            "type": "boolean"
          },
          "md5FailedError": {
            "type": [
              "null",
              "string"
            ]
          },
          "thumbnailsFailed": {
            "type": "boolean"
          },
          "thumbnailsFailedError": {
            "type": [
              "null",
              "string"
            ]
          },
          "thumbnailsGenerated": {
            "type": "boolean"
          },
          "importJobId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "fileSize": {
            "type": "integer",
            "format": "int64"
          },
          "duration": {
            "pattern": "^-?(\\d\u002B\\.)?\\d{2}:\\d{2}:\\d{2}(\\.\\d{1,7})?$",
            "type": "string"
          },
          "height": {
            "type": "integer",
            "format": "int32"
          },
          "width": {
            "type": "integer",
            "format": "int32"
          },
          "videoDimensionFormat": {
            "$ref": "#/components/schemas/VideoDimensionFormat"
          },
          "videoCodec": {
            "$ref": "#/components/schemas/VideoCodec"
          },
          "bitrate": {
            "type": "integer",
            "format": "int64"
          },
          "frameRate": {
            "type": "number",
            "format": "double"
          },
          "pixelFormat": {
            "type": [
              "null",
              "string"
            ]
          },
          "ratio": {
            "type": [
              "null",
              "string"
            ]
          },
          "creationTime": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "videoStreamCount": {
            "type": "integer",
            "format": "int32"
          },
          "audioStreamCount": {
            "type": "integer",
            "format": "int32"
          },
          "ingestDate": {
            "type": "string",
            "format": "date-time"
          },
          "cameraType": {
            "$ref": "#/components/schemas/CameraTypes"
          },
          "videoQuality": {
            "$ref": "#/components/schemas/VideoQuality"
          },
          "watchCount": {
            "type": "integer",
            "format": "int32"
          },
          "notes": {
            "type": "string"
          },
          "needsReview": {
            "type": "boolean"
          },
          "playbackIssue": {
            "type": "boolean"
          },
          "markedForDeletion": {
            "type": "boolean"
          },
          "isFavorite": {
            "type": "boolean"
          },
          "parentVideoId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "clipStartSeconds": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "clipEndSeconds": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "isClip": {
            "type": "boolean"
          },
          "isExportedClip": {
            "type": "boolean"
          },
          "isEdited": {
            "type": "boolean"
          },
          "chapterMarkers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChapterMarkerDto"
            }
          },
          "videoBlocks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VideoBlockDto"
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VideoTagSummaryDto"
            }
          },
          "properties": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PropertyValueDto"
            }
          }
        }
      },
...
    }
  },
...
}